requirements.txt
django-cors-middleware==1.3.1
djangorestframework==3.7.0
psycopg2==2.7.3.1
